Added registration list UI
Added a listview that displays an arraylist of temporary data. Still contains filler logs that should be replaced with actual code once the right features are implemented elsewhere, notably when we have a course details page to navigate to.

Note that *.xml had to be removed from the gitignore because we actually need the values in courseCategories.xml for the dropdown menu.

MR + MD