Refactoring & Small Bug Fix
BUG: When registered for a course, the listing does not appear initially as registered.

I renamed all classes to follow Camel-Case (with first letter capitalized), and changed a couple of class names to better reflect their contents (Registration -> CourseList)
The big refactoring was moving both User and Firebase singletons.
The user singleton is the currently logged in user, and the Firebase singleton is all the current firebase authentication and references.
I also collapsed firebase and firebaseDB into one class, and removed a duplicate Database class that was just mirroring firebaseDB.